% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balloon.R
\name{balloon_container}
\alias{balloon_container}
\alias{balloon}
\title{Ballons container and div}
\usage{
balloon_container(title, ..., title_tag = "h2")

balloon(..., side = c("left", "right"))
}
\arguments{
\item{title}{title of the `balloon_container`}

\item{...}{content of the element}

\item{title_tag}{tag level for `balloon_container`}

\item{side}{side for `balloon`}
}
\value{
a taglist
}
\description{
Ballons container and div
}
