% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{button}
\alias{button}
\alias{button_primary}
\alias{button_success}
\alias{button_warning}
\alias{button_error}
\title{Create a button}
\usage{
button(id, ...)

button_primary(id, ...)

button_success(id, ...)

button_warning(id, ...)

button_error(id, ...)
}
\arguments{
\item{id}{the idea of the button}

\item{...}{the content of the button}
}
\value{
a taglist
}
\description{
Create a button
}
