% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartridge.R
\name{cartridge}
\alias{cartridge}
\title{Create a new NES Themed Shiny App}
\usage{
cartridge(title = NULL, subtitle = NULL, ..., footer = NULL)
}
\arguments{
\item{title}{the title of the App}

\item{subtitle}{Substitle of the App}

\item{...}{content of the body}

\item{footer}{content of the footer.}
}
\value{
a tagList
}
\description{
Create a new NES Themed Shiny App
}
