% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{container}
\alias{container}
\alias{container_with_title}
\alias{container_simple}
\title{Customisable containers}
\usage{
container(..., is_dark = FALSE, is_rounded)

container_with_title(title = NULL, ..., is_centered = FALSE,
  is_dark = FALSE, title_tag = "h2")

container_simple(..., is_centered = FALSE, is_dark = FALSE,
  is_rounded = FALSE)
}
\arguments{
\item{...}{Content of the container}

\item{is_dark}{Should the container be dark?}

\item{is_rounded}{Should the container be rounded?}

\item{title}{Title of the container}

\item{is_centered}{Should the container be centered?}

\item{title_tag}{Header level of the title of the container}
}
\value{
a taglist
}
\description{
Customisable containers
}
