% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon.R
\name{heart}
\alias{heart}
\alias{star}
\alias{like}
\alias{twitter}
\alias{facebook}
\alias{github}
\alias{close}
\alias{octocat_animate}
\alias{nes_logo}
\alias{nes_jp_logo}
\alias{snes_logo}
\alias{snes_jp_logo}
\title{Various 8 bit icons}
\usage{
heart(size = c("small", "medium", "large"))

star(size = c("small", "medium", "large"))

like(size = c("small", "medium", "large"))

twitter(size = c("small", "medium", "large"))

facebook(size = c("small", "medium", "large"))

github(size = c("small", "medium", "large"))

close(size = c("small", "medium", "large"))

octocat_animate()

nes_logo()

nes_jp_logo()

snes_logo()

snes_jp_logo()
}
\arguments{
\item{size}{Wether to display in small, medium or large}
}
\value{
a taglist
}
\description{
Various 8 bit icons
}
