% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{nessy_examples}
\alias{nessy_examples}
\title{Get a NES example}
\usage{
nessy_examples(which = NULL)
}
\arguments{
\item{which}{The example to run. If empty, all the available examples are listed.}
}
\value{
A path to the example.
}
\description{
Get a NES example
}
\examples{
nessy_examples()
}
