% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{text_input}
\alias{text_input}
\title{Somewhere to enter text}
\usage{
text_input(id, label, value = "", width = NULL, type = c("none",
  "warning", "success", "error"), placeholder = NULL, inline = FALSE)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}

\item{type}{the style of the input}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{inline}{Should the label be inline?}
}
\value{
a tagList
}
\description{
Somewhere to enter text
}
